+++ docs/changelog.md
@@ -10,6 +10,7 @@ Added:
 - Add a `tags` attribute allowing cathegorizing `Variable`s (!354).
 - Add support for `complex` numbers (!374)
 - GUI: Add a Dark theme (!376)
+- add a`transient` attribute for variables we don't want to save data (!xxx)
 
 Changed: